Squashed 'external/NoctLang/' changes from 18675b4..1662b0c
1662b0c Improve translation
14cbdb3 Add -Wall flag
f0e1930 Improve translations
5cb6fe4 Merge NoctLang
1220ce7 Merge NoctLang
eb2712c Merge NoctLang
3405ce9 Merge NoctLang
4a29587 Merge NoctLang
c98bb6c Merge NoctLang
88330b3 Merge upstream NoctLang
3b1d485 Merge NoctLang
6ee85f1 Merge NoctLang
4d74f61 Merge NoctLang
7066b28 Fix debug build

git-subtree-dir: external/NoctLang
git-subtree-split: 1662b0c3ff165f8dd9f0f1cdef6511b3838266f2

+++ src/api/api-math.c
@@ -32,7 +32,7 @@ static bool cfunc_Math_random(struct rt_env *env);
 struct ffi_item {
 	const char *global_name;
 	const char *field_name;
-	int param_count;
+	uint32_t param_count;
 	const char *param[NOCT_ARG_MAX];
 	bool (*cfunc)(NoctEnv *env);
 };
@@ -63,7 +63,7 @@ bool noct_register_api_math(NoctEnv *env)
 		return false;
 
 	/* Register functions. */
-	for (i = 0; i < sizeof(ffi_items) / sizeof(struct ffi_item); i++) {
+	for (i = 0; i < (int)(sizeof(ffi_items) / sizeof(struct ffi_item)); i++) {
 		NoctValue funcval;
 
 		/* Register a cfunc. */
